Let's make the types used in the C callbacks slightly more consistent as well

+++ include/callbacks.h
@@ -11,6 +11,7 @@
  */
 
 #include <stddef.h>
+#include <stdint.h>
 
 /**
  *  Structure with all the callbacks
@@ -19,7 +20,7 @@ struct smart_tpl_callbacks {
     void        (*write)                (void *userdata, const char *data, size_t size);
     void        (*output)               (void *userdata, const void *variable, int escape);
     const void *(*member)               (void *userdata, const void *variable, const char *name, size_t size);
-    const void *(*member_at)            (void *userdata, const void *variable, long position);
+    const void *(*member_at)            (void *userdata, const void *variable, uint32_t position);
     void       *(*create_iterator)      (void *userdata, const void *variable);
     void        (*delete_iterator)      (void *userdata, void *iterator);
     int         (*valid_iterator)       (void *userdata, void *iterator);
@@ -28,19 +29,19 @@ struct smart_tpl_callbacks {
     void        (*iterator_next)        (void *userdata, void *iterator);
     const void *(*variable)             (void *userdata, const char *name, size_t size);
     const char *(*to_string)            (void *userdata, const void *variable);
-    int         (*to_numeric)           (void *userdata, const void *variable);
+    int64_t     (*to_numeric)           (void *userdata, const void *variable);
     int         (*to_boolean)           (void *userdata, const void *variable);
     size_t      (*size)                 (void *userdata, const void *variable);
     void       *(*modifier)             (void *userdata, const char *name, size_t size);
     const void *(*modify_variable)      (void *userdata, const void *variable, void *modifier, void *parameters);
-    void        (*assign_numeric)       (void *userdata, const char *key, size_t keysize, long value);
+    void        (*assign_numeric)       (void *userdata, const char *key, size_t keysize, int64_t value);
     void        (*assign_boolean)       (void *userdata, const char *key, size_t keysize, int boolean);
     void        (*assign_string)        (void *userdata, const char *key, size_t keysize, const char *buf, size_t buf_size);
     void        (*assign)               (void *userdata, const char *key, size_t keysize, const void *variable);
     int         (*strcmp)               (void *userdata, const char *a, size_t a_len, const char *b, size_t b_len);
     void       *(*create_params)        (void *userdata);
     void        (*delete_params)        (void *userdata, void *parameters);
-    void        (*params_append_numeric)(void *userdata, void *parameters, long value);
+    void        (*params_append_numeric)(void *userdata, void *parameters, int64_t value);
     void        (*params_append_string) (void *userdata, void *parameters, const char *buf, size_t len);
     void        (*params_append_boolean)(void *userdata, void *parameters, int boolean);
 };
